+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Request fiat on-ramp fees separately and use the returned `extraFee` as the MetaMask fee for standard fiat quotes. ([#9317](https://github.com/MetaMask/core/pull/9317))
 - Bump `@metamask/remote-feature-flag-controller` from `^5.0.0` to `^6.0.0` ([#9945](https://github.com/MetaMask/core/pull/9945))
 
 ## [26.4.1]